Move MatchNationalNumber into internal/regexbasedmatcher
MatchNationalNumber and its match helper move from the root package to
internal/regexbasedmatcher (mirroring upstream's internal/RegexBasedMatcher +
MatcherApi), taking *metadata.PhoneNumberDesc now that the proto types live in
the metadata package. Removes MatchNationalNumber from the public API.
Completes #5.
